slightly unbreak imports
