+++ src/ibex_bluesky_core/plans/__init__.py
@@ -1,4 +1,4 @@
-"""Generic plan helpers."""
+"""Generic plans."""
 
 from abc import ABC
 from collections.abc import Generator
@@ -16,8 +16,7 @@ from ibex_bluesky_core.callbacks import FitMethod, ISISCallbacks
 from ibex_bluesky_core.callbacks.fitting.fitting_utils import Linear
 from ibex_bluesky_core.devices import get_pv_prefix
 from ibex_bluesky_core.devices.block import BlockMot
-from ibex_bluesky_core.devices.dae import monitor_normalising_dae
-from ibex_bluesky_core.devices.simpledae import SimpleDae
+from ibex_bluesky_core.devices.simpledae import SimpleDae, monitor_normalising_dae
 from ibex_bluesky_core.plan_stubs import set_num_periods
 from ibex_bluesky_core.utils import centred_pixel
